// trade-service/src/common/enums/trade.enum.ts
export enum TradeType {
  EQUITY = 'EQUITY',
  OPTIONS = 'OPTIONS',
  FUTURES = 'FUTURES',
  FOREX = 'FOREX',
  CRYPTO = 'CRYPTO',
  COMMODITY = 'COMMODITY',
  INDEX = 'INDEX',
}

export enum Direction {
  LONG = 'LONG',
  SHORT = 'SHORT',
}

export enum TradeStatus {
  OPEN = 'OPEN',
  CLOSED = 'CLOSED',
  REVIEWED = 'REVIEWED',
}